requirements.txt
streamlit==1.31.0
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
pandas==2.0.3
spacy==3.7.2
python-dotenv==1.0.0
pyperclip
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
